autotracker/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.125.0
google-auth>=2.35.0
google-cloud-tasks>=2.16.0
google-cloud-secret-manager>=2.20.0
google-cloud-logging>=3.10.0
google-generativeai>=0.7.0
streamlit>=1.28.0
pandas>=2.0.0
redis>=5.0.0
pytest>=8.0.0
pydantic>=2.9.0
